+++ apps/quickbooks_online/exceptions.py
@@ -6,9 +6,8 @@ from fyle.platform.exceptions import InvalidTokenError as FyleInvalidTokenError
 from fyle_qbo_api.utils import invalidate_qbo_credentials
 from qbosdk.exceptions import InternalServerError, InvalidTokenError, WrongParamsError
 
-from apps.fyle.actions import update_failed_expenses
+from apps.fyle.actions import update_failed_expenses, post_accounting_export_summary
 from apps.fyle.models import ExpenseGroup
-from apps.fyle.tasks import post_accounting_export_summary
 from apps.quickbooks_online.actions import update_last_export_details
 from apps.quickbooks_online.errors.helpers import error_matcher, get_entity_values, replace_destination_id_with_values
 from apps.tasks.models import Error, TaskLog
