+++ scripts/python/post-accounting-export-summary.py
@@ -5,7 +5,7 @@ from django.db.models import Q
 
 from apps.workspaces.models import Workspace
 from apps.fyle.models import Expense
-from apps.fyle.tasks import post_accounting_export_summary
+from apps.fyle.actions import post_accounting_export_summary
 
 # PLEASE RUN scripts/python/fill-accounting-export-summary.py BEFORE RUNNING THIS SCRIPT
 workspaces = Workspace.objects.filter(
